requirements.txt
2404-segmentation-pipeline==0.1.2
monai==1.3.0
Flask==3.0.3